Adding test menus
