+++ packages/terra-framework-site/src/Tests.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 // import AppDelegate from 'terra-app-delegate';
-import { Route, Switch, Redirect, Link, withRouter } from 'react-router-dom';
+import { Route, Switch, Redirect, withRouter } from 'react-router-dom';
 
 import LayoutTests from 'terra-layout/tests/nightwatch/LayoutTests';
 import LayoutDefault from 'terra-layout/tests/nightwatch/LayoutDefault';
@@ -15,11 +15,9 @@ const Tests = () => (
         render={() => (
           <div style={{ height: '100%' }}>
             <Switch>
-              <Route path="/tests/layout-tests/default" component={LayoutDefault} />
-              <Route path="/tests/layout-tests/no-menu" component={LayoutNoMenu} />
-              <Route path="/tests/layout-tests/long-text" component={LayoutLongText} />
-              <Route path="/tests/layout-tests" component={LayoutTests} />
-              <Redirect to="/tests/layout-tests" />
+              <Route path="/tests/layout/default" component={LayoutDefault} />
+              <Route path="/tests/layout/no-menu" component={LayoutNoMenu} />
+              <Route path="/tests/layout/long-text" component={LayoutLongText} />
             </Switch>
           </div>
         )}
